pyproject.toml
[project]
name = "ontology-engine"
version = "0.1.0"
description = "Agentic Markdown-to-OWL ontology pipeline using Claude Agent SDK"
requires-python = ">=3.11"
dependencies = [
    "claude-agent-sdk>=0.1.34",
    "click",
    "python-dotenv",
    "rdflib",
    "rich",
]

[project.scripts]
ontology-engine = "ontology_engine.main:cli"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/ontology_engine"]
